Fixes dependencies and logout controller
This fixes the urls to successfully load dependencies through requirejs
using the config.js file. There is also a fix in the requirejs module
naming in the config.js file that was wrong (auth0-angular).
Also fixes an issue with the Logout controller to load the store as a
service.

Fixes https://github.com/auth0/auth0-angular/issues/236