Fix an error which happens when there is no error callback
auth0-angular requires not having callbacks set in order to use redirect mode. Not setting callbacks
causes the underlying auth0 library to throw an error. I'm changing
auth0-angular so it behaves like usual, but always sets an empty error callback
to the auth0 library, so auth0 does not throw an error.